more changes 24/25/7
