+++ kalima/kalima/doctype/applicant_student/applicant_student.py
@@ -20,7 +20,8 @@ class ApplicantStudent(Document):
             if age < 17:
                 frappe.throw(_("Age must be more than 17 years."))
 
-                
+    def before_save(self):
+        self.full_name_in_arabic = f"{self.first_name} {self.second_name} {self.third_name} {self.fourth_name}"
                 
 @frappe.whitelist()
 def admit_student(doc_name, department, study_system):
