+++ kalima/kalima/doctype/exam_questions/exam_questions.json
@@ -72,7 +72,7 @@
    "fieldtype": "Link",
    "in_list_view": 1,
    "label": "Department",
-   "options": "Presented Department",
+   "options": "Department",
    "read_only": 1
   },
   {
@@ -135,7 +135,7 @@
  "index_web_pages_for_search": 1,
  "is_submittable": 1,
  "links": [],
- "modified": "2024-07-06 15:42:02.653412",
+ "modified": "2024-07-25 10:22:09.898880",
  "modified_by": "Administrator",
  "module": "Kalima",
  "name": "Exam Questions",
