--- kalima/kalima/doctype/faculty_department/faculty_department.json
@@ -1,115 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "field:arabic_title",
- "creation": "2024-05-25 15:38:01.500302",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "faculty",
-  "english_title",
-  "arabic_title",
-  "abbreviation_title",
-  "code",
-  "column_break_qxzg",
-  "hr_department",
-  "tuition_fee",
-  "total_of_required_credits",
-  "passed_grade",
-  "is_active"
- ],
- "fields": [
-  {
-   "fieldname": "faculty",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Faculty",
-   "options": "Faculty",
-   "reqd": 1
-  },
-  {
-   "fetch_from": "faculty.english_title",
-   "fieldname": "english_title",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "English Title",
-   "reqd": 1
-  },
-  {
-   "fieldname": "arabic_title",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Arabic Title",
-   "reqd": 1,
-   "unique": 1
-  },
-  {
-   "fieldname": "abbreviation_title",
-   "fieldtype": "Data",
-   "label": "Abbreviation Title"
-  },
-  {
-   "fieldname": "code",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Code",
-   "reqd": 1
-  },
-  {
-   "fieldname": "column_break_qxzg",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "hr_department",
-   "fieldtype": "Link",
-   "label": "Hr Department",
-   "options": "Department"
-  },
-  {
-   "fieldname": "tuition_fee",
-   "fieldtype": "Currency",
-   "label": "Tuition Fee"
-  },
-  {
-   "fieldname": "total_of_required_credits",
-   "fieldtype": "Int",
-   "label": "Total of Required Credits"
-  },
-  {
-   "fieldname": "passed_grade",
-   "fieldtype": "Float",
-   "label": "Passed Grade"
-  },
-  {
-   "default": "1",
-   "fieldname": "is_active",
-   "fieldtype": "Check",
-   "label": "Is Active"
-  }
- ],
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2024-07-04 09:10:35.700828",
- "modified_by": "Administrator",
- "module": "Kalima",
- "name": "Faculty Department",
- "naming_rule": "By fieldname",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
