--- kalima/kalima/doctype/faculty_department/faculty_department.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, e2next and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class FacultyDepartment(Document):
-	pass
