--- kalima/kalima/doctype/group_class/group_class.json
@@ -1,154 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "format:KGRP-{title}-{#####}",
- "creation": "2024-05-26 11:18:54.340271",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "group_title",
-  "year",
-  "departments_groups",
-  "column_break_ibbt",
-  "stage",
-  "academic_system_type",
-  "semester",
-  "column_break_sglj",
-  "faculty",
-  "department",
-  "students_section",
-  "presented_modules",
-  "fetch_students",
-  "students",
-  "modules_html",
-  "students_html"
- ],
- "fields": [
-  {
-   "fieldname": "year",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Year",
-   "options": "Educational Year",
-   "reqd": 1
-  },
-  {
-   "fieldname": "stage",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Stage",
-   "options": "First Year\nSecond Year\nThird Year\nFourth Year\nFifth Year",
-   "reqd": 1
-  },
-  {
-   "fieldname": "semester",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Semester",
-   "options": "Fall Semester\nSpring Semester\nShort Semester\nAnnual",
-   "reqd": 1
-  },
-  {
-   "fieldname": "departments_groups",
-   "fieldtype": "HTML",
-   "label": "Departments Groups"
-  },
-  {
-   "fieldname": "group_title",
-   "fieldtype": "Data",
-   "label": "Group Title",
-   "reqd": 1,
-   "unique": 1
-  },
-  {
-   "fieldname": "column_break_ibbt",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "faculty",
-   "fieldtype": "Link",
-   "label": "Faculty",
-   "options": "Faculty",
-   "reqd": 1
-  },
-  {
-   "fieldname": "department",
-   "fieldtype": "Link",
-   "label": "Department",
-   "options": "Department",
-   "reqd": 1
-  },
-  {
-   "fieldname": "presented_modules",
-   "fieldtype": "HTML",
-   "label": "Presented Modules",
-   "read_only": 1
-  },
-  {
-   "fieldname": "students_section",
-   "fieldtype": "Section Break",
-   "label": "Students"
-  },
-  {
-   "fieldname": "fetch_students",
-   "fieldtype": "Button",
-   "label": "Fetch Students"
-  },
-  {
-   "fieldname": "students",
-   "fieldtype": "HTML",
-   "label": "Students"
-  },
-  {
-   "fieldname": "modules_html",
-   "fieldtype": "Text",
-   "hidden": 1,
-   "label": "modules_html",
-   "read_only": 1
-  },
-  {
-   "fieldname": "students_html",
-   "fieldtype": "Text",
-   "hidden": 1,
-   "label": "students_html",
-   "read_only": 1
-  },
-  {
-   "fieldname": "column_break_sglj",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "academic_system_type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Academic System Type",
-   "options": "Coursat\nBologna\nAnnual",
-   "reqd": 1
-  }
- ],
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2024-07-17 14:50:06.556628",
- "modified_by": "Administrator",
- "module": "Kalima",
- "name": "Group Class",
- "naming_rule": "Expression",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
