--- kalima/kalima/doctype/group_class/test_group_class.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, e2next and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestGroupClass(FrappeTestCase):
-	pass
