+++ kalima/kalima/doctype/outgoing/outgoing.json
@@ -1,7 +1,7 @@
 {
  "actions": [],
  "allow_rename": 1,
- "autoname": "format:KOUT-{date}-{#####}",
+ "autoname": "format:KOUT-{#####}",
  "creation": "2024-06-05 11:17:04.895201",
  "doctype": "DocType",
  "engine": "InnoDB",
@@ -67,7 +67,8 @@
   {
    "fieldname": "description",
    "fieldtype": "Text Editor",
-   "label": "Description"
+   "label": "Description",
+   "reqd": 1
   },
   {
    "fieldname": "receivers_type",
@@ -133,7 +134,7 @@
  "index_web_pages_for_search": 1,
  "is_submittable": 1,
  "links": [],
- "modified": "2024-07-24 13:24:17.786974",
+ "modified": "2024-07-24 16:39:10.959751",
  "modified_by": "Administrator",
  "module": "Kalima",
  "name": "Outgoing",
