--- kalima/kalima/doctype/presented_department/presented_department.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2024, e2next and contributors
-// For license information, please see license.txt
-
-// frappe.ui.form.on("Presented Department", {
-// 	refresh(frm) {
-
-// 	},
-// });
