--- kalima/kalima/doctype/presented_department/presented_department.json
@@ -1,118 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "format:KPRES-DEP-{#####}",
- "creation": "2024-05-26 11:07:50.620905",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "academic_semester",
-  "study_level",
-  "academic_system_type",
-  "entrance_type",
-  "max_available_semester",
-  "column_break_mmil",
-  "faculty",
-  "department",
-  "total_of_required_credits",
-  "total_of_years",
-  "tuition_fee"
- ],
- "fields": [
-  {
-   "fieldname": "academic_semester",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Academic Semester",
-   "options": "Fall Semester\nSpring Semester\nShort Semester\nAnnual",
-   "reqd": 1
-  },
-  {
-   "fieldname": "study_level",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Study Level",
-   "options": "Bachelor\nMaster\nPhd",
-   "reqd": 1
-  },
-  {
-   "fieldname": "academic_system_type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Academic System Type",
-   "options": "Coursat\nBologna\nAnnual",
-   "reqd": 1
-  },
-  {
-   "fieldname": "entrance_type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Entrance Type",
-   "options": "Morning\nEvening",
-   "reqd": 1
-  },
-  {
-   "fieldname": "max_available_semester",
-   "fieldtype": "Int",
-   "label": "Max Available Semester"
-  },
-  {
-   "fieldname": "column_break_mmil",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "faculty",
-   "fieldtype": "Link",
-   "label": "Faculty",
-   "options": "Faculty",
-   "reqd": 1
-  },
-  {
-   "fieldname": "department",
-   "fieldtype": "Link",
-   "label": "Department",
-   "options": "Department",
-   "reqd": 1
-  },
-  {
-   "fieldname": "total_of_required_credits",
-   "fieldtype": "Int",
-   "label": "Total of Required Credits"
-  },
-  {
-   "fieldname": "total_of_years",
-   "fieldtype": "Int",
-   "label": "Total of Years"
-  },
-  {
-   "fieldname": "tuition_fee",
-   "fieldtype": "Currency",
-   "label": "Tuition Fee"
-  }
- ],
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2024-06-10 13:31:21.518692",
- "modified_by": "Administrator",
- "module": "Kalima",
- "name": "Presented Department",
- "naming_rule": "Expression",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
