--- kalima/kalima/doctype/presented_department/test_presented_department.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, e2next and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestPresentedDepartment(FrappeTestCase):
-	pass
