--- kalima/kalima/doctype/presented_departments_table/presented_departments_table.json
@@ -1,102 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2024-05-29 14:57:45.157272",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "year",
-  "academic_semester",
-  "study_level",
-  "academic_system_type",
-  "status",
-  "entrance_type",
-  "max_available_semester",
-  "total_of_required_credits",
-  "total_of_years",
-  "tuition_fee"
- ],
- "fields": [
-  {
-   "fieldname": "academic_semester",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Academic Semester",
-   "options": "Fall Semester\nSpring Semester\nShort Semester\nAnnual",
-   "reqd": 1
-  },
-  {
-   "fieldname": "study_level",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Study Level",
-   "options": "Bachelor\nMaster\nPhd",
-   "reqd": 1
-  },
-  {
-   "fieldname": "academic_system_type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Academic System Type",
-   "options": "Coursat\nBologna\nAnnual",
-   "reqd": 1
-  },
-  {
-   "fieldname": "entrance_type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Entrance Type",
-   "options": "Morning\nEvening",
-   "reqd": 1
-  },
-  {
-   "fieldname": "max_available_semester",
-   "fieldtype": "Int",
-   "label": "Max Available Semester"
-  },
-  {
-   "fieldname": "total_of_required_credits",
-   "fieldtype": "Int",
-   "label": "Total of Required Credits"
-  },
-  {
-   "fieldname": "total_of_years",
-   "fieldtype": "Int",
-   "label": "Total of Years"
-  },
-  {
-   "fieldname": "tuition_fee",
-   "fieldtype": "Currency",
-   "label": "Tuition Fee"
-  },
-  {
-   "fieldname": "year",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Year",
-   "options": "Educational Year",
-   "reqd": 1
-  },
-  {
-   "fieldname": "status",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Status",
-   "options": "Active\nInactive",
-   "reqd": 1
-  }
- ],
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2024-05-29 15:07:17.526202",
- "modified_by": "Administrator",
- "module": "Kalima",
- "name": "Presented Departments Table",
- "owner": "Administrator",
- "permissions": [],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
