+++ kalima/kalima/doctype/presented_module/presented_module.json
@@ -21,7 +21,6 @@
   "swl",
   "sswl",
   "usswl",
-  "available_in_choose_credit_panel",
   "scheduled_tab",
   "a_section",
   "in_class_hours",
@@ -122,6 +121,7 @@
    "reqd": 1
   },
   {
+   "depends_on": "eval:doc.academic_system_type == \"Bologna\"",
    "fieldname": "sswl",
    "fieldtype": "Int",
    "label": "SSWL",
@@ -135,16 +135,10 @@
    "read_only": 1
   },
   {
+   "depends_on": "eval:doc.academic_system_type == \"Bologna\"",
    "fieldname": "ects",
    "fieldtype": "Int",
    "label": "ECTS"
-  },
-  {
-   "default": "0",
-   "fieldname": "available_in_choose_credit_panel",
-   "fieldtype": "Check",
-   "label": "Available in Choose Credit Panel",
-   "options": "Module"
   },
   {
    "fieldname": "main_tab",
@@ -152,7 +146,7 @@
    "label": "Main"
   },
   {
-   "depends_on": "eval:doc.stage == \"Bologna\"",
+   "depends_on": "eval:doc.academic_system_type == \"Bologna\"",
    "fieldname": "scheduled_tab",
    "fieldtype": "Tab Break",
    "label": "Scheduled"
@@ -324,7 +318,7 @@
    "fieldtype": "Column Break"
   },
   {
-   "depends_on": "eval:doc.stage == \"Bologna\"",
+   "depends_on": "eval:doc.academic_system_type == \"Bologna\"",
    "fieldname": "unscheduled_tab",
    "fieldtype": "Tab Break",
    "label": "Unscheduled"
@@ -422,6 +416,7 @@
    "label": "Total Hours"
   },
   {
+   "depends_on": "eval:doc.academic_system_type == \"Bologna\"",
    "fieldname": "swl",
    "fieldtype": "Int",
    "label": "SWL",
@@ -442,7 +437,7 @@
  ],
  "index_web_pages_for_search": 1,
  "links": [],
- "modified": "2024-07-20 16:41:50.456644",
+ "modified": "2024-07-25 10:59:34.971582",
  "modified_by": "Administrator",
  "module": "Kalima",
  "name": "Presented Module",
