+++ kalima/kalima/doctype/student/student.py
@@ -56,7 +56,9 @@ class Student(Document):
 	
 			doc.total = ttl
 			doc.final_average = ttl/len(doc.ministry_exam_results)
-			
+   
+		doc.full_name_in_arabic = f"{doc.first_name} {doc.second_name} {doc.third_name} {doc.fourth_name}"
+                
 	def after_save(doc):
 		# user = frappe.get_doc("User", doc.user)
 		if(doc.password != None and doc.password != ""):
